chore: refactor the conversion of structureChecks to TreeNodes to the Project.ts file
